# [CRT] Improve rand_s