--- a/modules/rostests/apitests/crt/static_crt_apitest.cmake
+++ b/modules/rostests/apitests/crt/static_crt_apitest.cmake
@@ -13,6 +13,7 @@ list(APPEND SOURCE_STATIC
     fpcontrol.c
     mbstowcs.c
     mbtowc.c
+    rand_s.c
     sprintf.c
     strcpy.c
     strlen.c