--- a/sdk/lib/crt/msvcrtex.cmake
+++ b/sdk/lib/crt/msvcrtex.cmake
@@ -11,6 +11,13 @@ list(APPEND MSVCRTEX_SOURCE
     misc/ofmt_stub.c
     stdio/acrt_iob_func.c)
 
+if(DLL_EXPORT_VERSION LESS 0x600)
+    list(APPEND MSVCRTEX_SOURCE
+        stdlib/_invalid_parameter.c
+        stdlib/rand_s.c
+    )
+endif()
+
 if(CMAKE_C_COMPILER_ID STREQUAL "Clang")
     # Clang performs some optimizations requiring those funtions
     list(APPEND MSVCRTEX_SOURCE