… sdk/lib/crt/stdlib/rand.c
… sdk/lib/crt/stdlib/rand_nt.c