## Purpose

This improves rand_s to avoid the massive overhead of loading advapi32, opening a handle to ksecdd, closing the handle and unloading advapi32 for every single call to rand_s. Instead the function pointer is cached. Testing resulted in a speedup of 10,000 times when run on Windows 10 using the statically linked apitest.
This also adds rand_s to msvcrtex for linking with newer GCC stdc++.

## Tests
- Test WHS: https://reactos.org/testman/compare.php?ids=94813,94812 (static_crt:rand_s)
before: 5.94s - `rand_s took 13040596278 cycles`
after: 0.02s - `rand_s took 11622704 cycles`
=> 1000x

- KVM x86: https://reactos.org/testman/compare.php?ids=94818,94824
before: 0.51s - `rand_s took 1070359452 cycles`
after: 0.15s  `rand_s took 285251108 cycles`
=> 3.7x

This is a quite unexpected result, showing that our totally broken and in no way cryptographically secure user mode implementation of RtlGenRandom is actually 25x slower than the kernel mode based one in win 2003, while the overhead of loading advapi32 and opening the handle to ksecdd is somehow 12 times faster on reactos.