@JoachimHenze 2024-04-17
Did you have any specific testcase, which made you aware of the overhead? (e.g. by poor performance in a specific app?)