@tkreuzer 2024-04-19
> Did you have any specific testcase, which made you aware of the overhead? (e.g. by poor performance in a specific app?)

I wrote a test and that showed that the new implementations isn't "compatible" with Windows, so you cannot run it as a unittest. I decided to instead use the previous code, but instead of loading and unloading advapi32 for every call, the function pointer is cached. This speeds up the statically linked performance test by a factor of 10,000 on Windows 10. Our testbots will show what the benefit is on windows 2003 / reactos.