@RaulKong898 2024-04-20
Hi. I would like ReactOS to look and feel more like Windows XP, especially the system apps to be more like Windows XP. I also have the Windows XP source code that was leaked on the internet. This is the source code link: https://github.com/tongzx/nt5src. I really want ReactOS to look like Windows XP, especially the system apps, because Windows XP is no longer supported and we can easily get viruses and stuff. So with ReactOS we can modify the source code of Windows XP and add interesting things that never existed on Windows XP and we can have security support. Consider this message, especially since many people who lived with Windows XP would like to go back to Windows XP because it seemed simpler and different, especially for retro tech people and for older PCs that they are no longer used and if they are thrown away we destroy the planet if they are not recycled, especially since this way ReactOS will have support for many drivers from Windows XP.